+++ couriersync-backend/src/main/java/com/ep18/couriersync/backend/customers/api/graphql/DepartamentoGraphQL.java
@@ -1,10 +1,14 @@
 package com.ep18.couriersync.backend.customers.api.graphql;
 
 import com.ep18.couriersync.backend.common.dto.PagingDTOs.PageResponse;
-import com.ep18.couriersync.backend.customers.dto.DepartamentoDTOs.*;
+import com.ep18.couriersync.backend.customers.dto.DepartamentoDTOs.CreateDepartamentoInput;
+import com.ep18.couriersync.backend.customers.dto.DepartamentoDTOs.DepartamentoView;
+import com.ep18.couriersync.backend.customers.dto.DepartamentoDTOs.UpdateDepartamentoInput;
 import com.ep18.couriersync.backend.customers.service.DepartamentoService;
 import lombok.RequiredArgsConstructor;
-import org.springframework.graphql.data.method.annotation.*;
+import org.springframework.graphql.data.method.annotation.Argument;
+import org.springframework.graphql.data.method.annotation.MutationMapping;
+import org.springframework.graphql.data.method.annotation.QueryMapping;
 import org.springframework.stereotype.Controller;
 
 @Controller
@@ -14,7 +18,7 @@ public class DepartamentoGraphQL {
     private final DepartamentoService service;
 
     @QueryMapping
-    public DepartamentoView departamentoById(@Argument Long id) {
+    public DepartamentoView departamentoById(@Argument Integer id) {
         return service.findById(id);
     }
 
@@ -35,7 +39,7 @@ public class DepartamentoGraphQL {
     }
 
     @MutationMapping
-    public Boolean deleteDepartamento(@Argument Long id) {
+    public Boolean deleteDepartamento(@Argument Integer id) {
         return service.delete(id);
     }
 }
