+++ couriersync-backend/src/main/java/com/ep18/couriersync/backend/customers/repository/DepartamentoRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface DepartamentoRepository extends JpaRepository<Departamento, Long> {
+public interface DepartamentoRepository extends JpaRepository<Departamento, Integer> {
 
     Optional<Departamento> findByNombreDepartamentoIgnoreCase(String nombreDepartamento);
